requirements.txt
Flask==0.10.0
Markdown==2.5.2
google-api-python-client